+++ vercel.json
@@ -1,0 +1,15 @@
+{
+    "version": 2,
+    "builds": [
+        {
+            "src":"/app.py",
+            "use": "@vercel/python"
+        }
+    ],
+    "routes": [
+        {
+            "src":"/(.*)",
+            "dest": "app.py"
+        }
+    ]
+}
